No need to specify the file evry time you call a method
Every time you needed to call a method you need to specify the file which was relative to the original filepath you entered upon object instanciation. This system was overly complicated and over-engineered, so this update makes it so you don't need to specify the filepath every time you call a method and the filapath is no longer relative to some other path. You can change the object file path at run time using the `setFilepath` or get the current filepath at runtime using the `getFilepath` method.